requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
kubernetes>=29.0.0
websockets>=12.0
httpx>=0.27.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
